Implement pytest based test system
Automated tests are necessary to ensure a high code quality and program
reliability. The pytest framework was chosen due to its flexibility,
minimalism and widespread usage.

To run the tests, simply execute `pytest` in the root directory of the
repository. The setup.cfg file prevents the old manual tests from being
executed. The conftest.py file ensures that the repository version of
hdtv is being used instead of a system-wide installation.
helpers/utils.py provides a way to capture the stdout/stderr-output of hdtv.

As a first step, test were added for several (yet not all) plugins,
ensuring the operability of the recent argparse switch.

Possibilities to test the GUI were neither implemented nor explored.