Added man page, bash/zsh completion, .desktop-file
A man page was added providing a short description of hdtv. It is still
rather incomplete and should be expanded.
Additionally, a secondary manpage `hdtv-tutorial` should be added that
provides an introduction to HDTV.

A .desktop-file was added.

Completion files for bash and zsh are now provided. No completions are
currently possible for hdtv commands used in --execute.