Proper support for integrals, fitxml v1.4
Integrals are now properly supported in HDTV. An integral is a property
that is attached to every HDTV fit (or multiple properties for the
region integral, background integral and the difference).
The integral of the workFit is displayed using the I key or the command
"fit integral execute". To list the integrals of all fits, use
"fit integral list".
If one wants to work excleusively with integrals, it is possible to work
with peakless fits.

The xml fit files also (always) include information about the integrals.
The fitxml version was bumped to v1.4 accordingly.