Fix dlmgr usrlibdir path
The dlmgr uslibdir path for rebuilt libraries was changed to
reflect the recent (72238ed1, 0fb61cb3) addition of XDG base
directory support.

As the rebuilt libraries in this folder are easily recreated,
it was placed in the $XDG_CACHE_HOME. Backwards compatibility
for the legacy ~/.hdtv folder is not necessary.