Support for compressed .xfl files
Seamless compression of xml fitlists is now supported. If the filename
ends in '.xz', '.gz' or '.bz2', compression is automatically added both
for reading and writing. The user benefits from reduced storage space
and, in case of xz, internal checksumming.

In case of python2, xz and bz2 compression are not supported. The user
is warned about this. Compressed files can still be accessed using the
xz and bzip2 command line utilities.

A new option 'default_fitlist_extension' is introduced to automatically
write compressed fitlists.