Add check and warning for ROOT 6.22/00, 6.22/02
TPyDispatcher has been reintroduced to ROOT:
https://github.com/root-project/root/issues/6579

I tested the ROOT 6.23/00 nightly build, and hdtv is working correctly
again (tests are also passing). The fixes were backported to ROOT 6.22
and will be available with 6.22/04. I added this information to the
README, and because no one reads the README, I also added a version
check when starting HDTV.