calbin function and unnecessary print (#24)
* calbin command for negative-energy bins

If the original spectrum had bins with negative energy values, the calbin command would fail because this case was not accounted for (negative index in numpy-array access). Since calbin-ing assumes that the energies start at 0, negative-energy bins are now discarded and a warning message is printed.

* Remove unnecessary print command

Found that "fit read" always printed a list of data when executed. The print command was probably a debug output once.