Fixed wrong sorting of residual tables
This commit fixes a wrong sorting of the residual tables by channel
that can be printed by various calibration commands.
Before this commit for example the command
calibration position enter -t 0 0 10 10 20 20 5 5 200 200
printed the residual table with the channel sequence being 200, 10, 20,
0, 5. I.e. the channels were first sorted by order of magnitude and only
then by absolute magnitude.
(This was due to leading whitespace in the number strings being
considered for sorting as well)