Config option to sound bell when wrapping ID by next/prev
This commit adds a config option ui.bellOnNextAndPrevIDEndReached to
enable emission of the bell character \a to the terminal, whenever the
result of the special IDs next and prev is the first or last ID again
due to wrap around of the ID range.
This affects any type of ID parsing, i.e., including spectra IDs and
fit IDs.
Enabling this option will for example sound the bell when going through
spectra by repeatedly issuing the command spectrum show next (or
repeatedly pressing the page down key) as soon as one ends up with the
first spectrum again, hence, noticeable informing of this fact.