Added fit option onlypositivepeaks to Theuerkauf peak model
Add a fit parameter option onlypositivepeaks to the Theuerkauf peak
model, if set to true the underlying ROOT fitter will be told to only
allow positive peak widths and volumes in the fit.
Furthermore the initial fit value for the width is now set up to
always be positive (regardless of the new option) to avoid the ROOT
fitter to fit regular peaks with negative widths and volumes, this
could happen before in certain cases with low statistics peaks.