Use scikit-build-core
Remove the custom code to compile the Cern ROOT plugins during runtime
and instead use scikit-build-core to build them during installation.

On a Cern ROOT update just reinstall hdtv.

The benifits are much cleaner development process and skipping the need
to handle the compilation by the program logic itself. It also seperates
better into installation / compilation and runtime errors.

Additionally this is a preperation to bring hdtv to other platforms.